# docs/api/openapi-internal.yaml
openapi: 3.1.0
info:
  title: BetTrack Dashboard REST API
  description: |
    Internal REST API for the BetTrack sports betting dashboard backend.

    ## Authentication

    The API supports two modes controlled by the `AUTH_MODE` environment variable:

    - **`none`** (default) — All routes accessible without login. Admin routes still require a valid admin session.
    - **`oauth2`** — Session-based OAuth2 authentication via Google or Microsoft (Azure AD).

    MCP integration routes and AI bet creation use API key Bearer tokens prefixed with `sk_`.

    ## Base URL

    Default: `http://localhost:3001`

    All API routes are prefixed with `/api`.
  version: 0.2.11
  contact:
    url: https://github.com/WFord26/BetTrack
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://{domain}
    description: Production
    variables:
      domain:
        default: yourdomain.com
        description: Your production domain
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: bettrack_session
      description: Session cookie established after OAuth2 login. Not required when AUTH_MODE=none.
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: sk_<key>
      description: API key with `sk_` prefix. Generate via POST /api/keys.
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          description: Response payload (schema varies by endpoint)
...
    Game:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
        sportKey:
          type: string
          example: americanfootball_nfl
        sportName:
          type: string
          example: NFL
        homeTeamName:
          type: string
        awayTeamName:
          type: string
        commenceTime:
          type: string
          format: date-time
        venue:
          type: string
          nullable: true
        weather:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, in_progress, final, postponed]
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        period:
          type: string
          nullable: true
        clock:
          type: string
          nullable: true
        bookmakers:
          type: array
          description: Grouped odds by bookmaker
          items:
            type: object
            properties:
              key:
                type: string
              title:
                type: string
              markets:
                type: array
                items:
                  type: object
...
paths:
...
  /api/games:
    get:
      tags: [Games]
      summary: List games with filters
      description: Returns games and their current odds. Supports timezone-aware date filtering.
      security:
        - sessionAuth: []
      parameters:
        - name: date
          in: query
          description: Filter by user's local date (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2026-01-09"
        - name: sport
          in: query
          description: Sport key filter (e.g. `americanfootball_nfl`). Use `all` for all sports.
          schema:
            type: string
        - name: status
          in: query
          description: Game status filter
          schema:
            type: string
            enum: [scheduled, in_progress, final]
        - name: timezoneOffset
          in: query
          description: Client timezone offset in minutes (from `Date.getTimezoneOffset()`)
          schema:
            type: integer
            example: 420
        - name: bookmaker
          in: query
          description: Filter odds by specific bookmaker key
          schema:
            type: string
      responses:
        '200':
          description: List of games with odds
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      games:
                        type: array
                        items:
                          $ref: '#/components/schemas/Game'
                      count:
                        type: integer
  /api/games/{id}:
    get:
      tags: [Games]
      summary: Get a single game with full odds detail
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game with full odds data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
        '404':
          description: Game not found
  /api/games/{id}/odds:
    get:
      tags: [Games]
      summary: Get current odds for a game
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: bookmaker
          in: query
          schema:
            type: string
        - name: marketType
          in: query
          description: Filter by market type
          schema:
            type: string
            enum: [h2h, spreads, totals]
      responses:
        '200':
          description: Current odds for the game
  /api/games/{id}/odds/history:
    get:
      tags: [Games]
      summary: Get odds movement history for a game
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: bookmaker
          in: query
          schema:
            type: string
        - name: marketType
          in: query
          schema:
            type: string
            enum: [h2h, spreads, totals]
        - name: hours
          in: query
          description: How many hours of history to return (default 24)
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Odds history snapshots sorted by time

  # ============================================================
  # BETS
  # ============================================================
...
  /api/stats/game/{gameId}:
    get:
      tags: [Stats]
      summary: Get team and player stats for a game
      description: Returns box score stats, season averages, and player game logs for both teams.
      security:
        - sessionAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game stats with team and player breakdowns
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      teamStats:
                        type: array
                        items:
                          type: object
                      playerStats:
                        type: array
                        items:
                          type: object
                      seasonAverages:
                        type: array
                        items:
                          type: object
        '404':
          description: Game not found
...